create-simple-integration-test
The next step is to create a simple integration test using Capybara.
There are two parts to this: the magic boilerplate and the meat of the
test.

For the magic boiler plate, a quick Google gives us [this
post](http://blag.ahax.de/post/1581758817/using-capybara-with-plain-rails-integration-tests-and).

For the meat of the test, we refer to the [capybara
documentation](https://github.com/jnicklas/capybara/blob/master/README.rdoc).
Look at the section titled `The DSL`, which contains links to
appropriate sections in the rdoc.

SHOW_PATCH

When your test is written, run it with

    $ rake test:integration